Move JSON schema helpers to json ingestion gem
