+++ elasticgraph-schema_definition/README.md
@@ -21,6 +21,9 @@ graph LR;
     elasticgraph-indexer["elasticgraph-indexer"];
     elasticgraph-schema_definition --> elasticgraph-indexer;
     class elasticgraph-indexer otherEgGemStyle;
+    elasticgraph-json_ingestion["elasticgraph-json_ingestion"];
+    elasticgraph-schema_definition --> elasticgraph-json_ingestion;
+    class elasticgraph-json_ingestion otherEgGemStyle;
     elasticgraph-schema_artifacts["elasticgraph-schema_artifacts"];
     elasticgraph-schema_definition --> elasticgraph-schema_artifacts;
     class elasticgraph-schema_artifacts otherEgGemStyle;
