+++ elasticgraph/lib/elastic_graph/project_template/Gemfile.tt
@@ -7,6 +7,7 @@ gem "graphql-c_parser", "~> 1.1", platforms: :ruby
 elasticgraph_details = <%= ElasticGraph.setup_env.gemfile_elasticgraph_details_code_snippet %>
 
 gem "elasticgraph-local", *elasticgraph_details
+gem "elasticgraph-json_ingestion", *elasticgraph_details
 gem "elasticgraph-query_registry", *elasticgraph_details
 
 # Can be elasticgraph-elasticsearch or elasticgraph-opensearch based on the datastore you want to use.
